+++ src/bundler.ts
@@ -188,8 +188,9 @@ export class Bundler {
       bundleImports?: Set<string>): Promise<ASTNode> {
     const docUrl = docBundle.url;
     const document = await this._prepareBundleDocument(docBundle);
+
     this._appendHtmlImportsForBundle(document, docBundle);
-    importUtils.rewriteImportedUrls(document, docUrl, docUrl);
+    importUtils.rewriteDocumentToEmulateBaseTag(docUrl, document);
 
     await this._inlineHtmlImports(docUrl, document, docBundle, bundleManifest);